refactor : dto 객체 default 생성자로 수정 및 리패키징, 및 도메인에서 잘 못된 의존관계 삭제(pieceDto)
