+++ src/main/java/chess/controller/room/RoomRestController.java
@@ -1,4 +1,4 @@
-package chess.controller;
+package chess.controller.room;
 
 import chess.chessgame.domain.room.game.ChessGameManager;
 import chess.chessgame.domain.room.user.User;
